+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- `Innmind\Specification\Comparator\Property`
+
 ## 4.0.0 - 2024-07-06
 
 ### Removed